+++ custom_components/pp_reader/www/pp_reader_dashboard/js/dashboard.module.js
@@ -1,1 +1,1 @@
-export * from './dashboard.BpIopZ75.js';
+export * from './dashboard.DCZyrOlU.js';